Remove rez.utils.py23 and cleanup more python 2.7 leftovers (#1678)
* remove rez.utils.py23 module
* remove some more __nonzero__ methods

---------
